Added balanced class-splitting functionality to parse_labels
Added class splitting option to parse_labels and generated a new set of labels using the default settings.
This conforms better to a subtle point of prototypical networks that was not clear before. Performance and actual
rationale are pending, but it seems to provide more order in this ludacris dataset